refactor: getter를 개선하기 위해 match의 책임을 LotteryNumbers로 이동하고 WinnerLottery에서 호출하도록 수정
